+++ packages/nimbus/src/components/form-field/components/form-field.info-box.tsx
@@ -1,7 +1,14 @@
 import { useContext, useEffect, type ReactNode } from "react";
-import { FormFieldContext } from "../form-field";
+import { FormFieldContext } from "./form-field.context";
 
-export const FormFieldInfoBox = ({ children }: { children: ReactNode }) => {
+type FormFieldInfoBoxProps = {
+  /**
+   * The content to display in the InfoBox
+   */
+  children: ReactNode;
+};
+
+export const FormFieldInfoBox = ({ children }: FormFieldInfoBoxProps) => {
   const { setContext } = useContext(FormFieldContext);
 
   useEffect(() => {
