+++ packages/nimbus/src/components/form-field/components/form-field.label.tsx
@@ -1,5 +1,5 @@
 import { useContext, useEffect } from "react";
-import { FormFieldContext } from "../form-field";
+import { FormFieldContext } from "./form-field.context";
 import type { FormFieldLabelSlotProps } from "../form-field.slots";
 
 export const FormFieldLabel = ({
